fix: resolve Epic #367 code-review violations
- Replace recv_timeout(100ms) polling with recv() in configure_cmake
  and build_project so no log lines are lost during subprocess silence

- Remove 4 hardcoded -DCMAKE_CXX_FLAGS / -DCMAKE_C_FLAGS args that
  silently overrode the merge_flags() env-var merging

- Evacuate CLI rendering from gglib-runtime: move handle_install,
  consume_build_events_cli, build_from_source_impl,
  determine_acceleration, and print_preflight_info into new
  crates/gglib-cli/src/handlers/llama_install.rs; move indicatif
  dependency to gglib-cli; wire dispatch in handlers/llama.rs

- ensure.rs: replace handle_install calls with local install_from_source
  helper that calls run_llama_source_build directly with plain println
  output (no indicatif) so gglib-runtime stays surface-agnostic

- Fix merge_flags doc (was "Parse build progress from CMake output")

- Delete ## Note block from build/README.md (claimed CXXFLAGS override
  was safe; now factually wrong after cmake args removed)
