+++ crates/gglib-runtime/Cargo.toml
@@ -50,7 +50,7 @@ zip = { version = "7.1", default-features = false, features = ["aes-crypto", "bz
 flate2 = "1"
 tar = "0.4"
 
-# CLI progress (used by cli feature)
+# CLI progress (used by cli feature and prebuilt download)
 indicatif = { workspace = true }
 
 # Unix signal handling
