+++ crates/gglib-runtime/src/llama/mod.rs
@@ -63,6 +63,9 @@ pub use prompt::{AutoConfirmPrompt, InstallPrompt, NonInteractivePrompt};
 pub use build_events::{BuildEvent, BuildPhase};
 
 #[cfg(feature = "cli")]
+pub use deps::{check_dependencies, check_disk_space};
+
+#[cfg(feature = "cli")]
 pub use progress::CliProgress;
 
 #[cfg(feature = "cli")]
@@ -77,11 +80,9 @@ pub use validate::{handle_status, validate_llama_binary, validate_llama_cli_bina
 
 // Installation (CLI only)
 #[cfg(feature = "cli")]
-pub use install::handle_install;
-#[cfg(feature = "cli")]
 pub use install::run_llama_source_build;
 #[cfg(feature = "cli")]
-pub use uninstall::{handle_rebuild, handle_uninstall};
+pub use uninstall::handle_uninstall;
 #[cfg(feature = "cli")]
 pub use update::{handle_check_updates, handle_update};
